mesa: Set `soc_specific: true` for installed targets
Hardware drivers must be installed in /vendor on Android, and
`soc_specific: true` causes targets to be installed there.

It also appears to change the set of headers available, which is why
this patch also adds `libnativebase_headers` to libEGL. Otherwise it
fails to build because it cannot `#include <nativebase/nativebase.h>`.

To avoid needing to add other `header_libs` to intermediate static
libraries, only set `soc_specific: true` on installed targets.