add project method to filter input targets
When it comes to whole static libraries, meson can sometimes generates
`-Wl,--whole-archive`, which is supported, but it can also just link
the library with the static libraries.

This can cause issues if 2 whole static libraries define the same
variable, as it would end up being defined twice (only value would
really be kept by the compiler).

Allow to filter out input targets to make sure to compile them
separately.

This is more a workaround than a fix. A proper way to deal with that
would be to compare the properties of each input before merging them
in a library source.