Disable wildcard in CI (#185)
The goal is to remove the need for checkouting directories when using
the CI cache.

Also change the behavior of `--copy-to-aosp`. When used `Android.bp`
is directly copied to AOSP, without being copied in ninja-to-soong.

No copied is made if there is no change detected (simple comparison).

Add a unittest for wildcarding.